[FE] refactor: setting 관련 interface 및 파일 분리
