+++ frontend/src/interfaces/setting.interface.ts
@@ -15,6 +15,11 @@ export interface KeyConfig {
   id: number;
 }
 
+export interface ComponentConfig {
+  key: string;
+  id: string;
+}
+
 export interface EditListConfig extends KeyConfig {
   editable: boolean;
 }
@@ -26,4 +31,36 @@ export interface ComponentCheckConfig {
   page?: boolean;
   title?: boolean;
   contents: boolean;
+}
+
+export interface ColorsConfig {
+  title: {
+    background: string;
+    text: string;
+    titleHeight?: number;
+    type: boolean; // 이미지 or 색
+    titleText: string;
+  };
+  category: {
+    background: string;
+    text: string;
+  };
+  page: {
+    background: string;
+    text: string;
+    sort: string;
+  };
+  profile: {
+    background: string;
+    text: string;
+  };
+  contents: {
+    background: string;
+    text: string;
+  };
+  logo: {
+    background: string;
+    text: string;
+    logoText: string;
+  };
 }
